Reorganization of files
